Changed int to doubles.

+++ app/src/main/java/blankthings/raincheckmvp/net/data/Coord.kt
@@ -5,9 +5,9 @@ import com.google.gson.annotations.SerializedName
 class Coord {
 
     @SerializedName("lat")
-    var lat: Double = 0.toDouble()
+    var lat: Double = 0.0
 
     @SerializedName("lon")
-    var lon: Double = 0.toDouble()
+    var lon: Double = 0.0
 
 }
