+++ app/src/main/java/blankthings/raincheckmvp/net/data/Main.kt
@@ -5,27 +5,27 @@ import com.google.gson.annotations.SerializedName
 class Main {
 
     @SerializedName("temp")
-    var temp: Double = 0.toDouble()
+    var temp: Double = 0.0
 
     @SerializedName("temp_min")
-    var tempMin: Double = 0.toDouble()
+    var tempMin: Double = 0.0
 
     @SerializedName("temp_max")
-    var tempMax: Double = 0.toDouble()
+    var tempMax: Double = 0.0
 
     @SerializedName("pressure")
-    var pressure: Double = 0.toDouble()
+    var pressure: Double = 0.0
 
     @SerializedName("sea_level")
-    var seaLevel: Double = 0.toDouble()
+    var seaLevel: Double = 0.0
 
     @SerializedName("grnd_level")
-    var grndLevel: Double = 0.toDouble()
+    var grndLevel: Double = 0.0
 
     @SerializedName("humidity")
-    var humidity: Int = 0
+    var humidity: Double = 0.0
 
     @SerializedName("temp_kf")
-    var tempKf: Int = 0
+    var tempKf: Double = 0.0
 
 }
