+++ app/src/main/java/blankthings/raincheckmvp/net/data/Snow.kt
@@ -5,6 +5,6 @@ import com.google.gson.annotations.SerializedName
 class Snow {
 
     @SerializedName("3h")
-    var last3HourVolume: Double = 0.toDouble()
+    var last3HourVolume: Double = 0.0
 
 }
